Refactor: Clearer grouping name for ArsenalTests
