+++ WarGames.Business.NUnit/ArsenalTests/ValueTargettingTests.cs
@@ -11,7 +11,7 @@ using WarGames.Contracts.Game;
 using WarGames.Resources.Arsenal;
 using WarGames.Resources.Game;
 
-namespace WarGames.Business.NUnit.TargetingTests
+namespace WarGames.Business.NUnit.ArsenalTests
 {
 	[TestFixture]
 	public class ValueTargettingTests
